read-stream-content-into-string element-type fix
Default the element type of the string being created from the element
type of the stream being read.

This has two advantages: it will make things work for implementations
which have a mode where the default string element type is a subtype
of character, and it will also mean that if the stream's element type
is a subtype of character you may get thinner strings.