Fix issue 30 - testing value of `hash-table-size`.
As far as I can tell from the spec, the meaning of `hash-table-size`
is not dictated, and implementations are free to return just about
anything; it's just a way to provide useful information (hints) when
creating a new hash-table.

I disabled this test on Allegro originally, but Attila Lendvai
proposes we simply remove it altogether, as this commit does.